pyproject.toml
[tool.poetry]
name = "enochian_language_modeling"
version = "0.1.8"
description = "Small-corpora language modeling with a particular focus on the Enochian language—and some puzzles related to its corpora."
authors = ["Forrest Wilbur <forrest.a.wilbur@gmail.com>"]
readme = "README.md"
packages = [
  { include = "enochian_lm", from = "src" },
  { include = "loagaeth", from = "src" },
  { include = "translation", from = "src" },
  { include = "training", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
python-dotenv = "^1.1.0"
openai = ">=2.0,<3"
gensim = "^4.3.3"
scikit-learn = "^1.6.1"
python-levenshtein = "^0.27.1"
sentence-transformers = "^4.0.2"
tqdm = "^4.67.1"
tenacity = "^9.1.2"
networkx = "^3.5"
scikit-fuzzy = "^0.5.0"
marisa-trie = "^1.2.1"
hyperopt = "^0.2.7"
yaspin = "^3.1.0"
pysqlite3-binary = "^0.5.4"
pydantic = ">=2.8,<3"
pysqlite3 = "^0.5.4"
crewai = "^1.4.1"
numpy = "^2.3.4"
matplotlib = "^3.10.7"
pandas = "^2.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
enochian-build-fasttext = "enochian_lm.root_extraction.tools.train_fasttext_model:main"
enochian-build-ngram-index = "enochian_lm.root_extraction.utils.build_ngram_index:main"
enochian-inspect = "enochian_lm.root_extraction.tools.inspect_definitions:main"
enochian-analysis = "enochian_lm.root_extraction.main:main"
enochian-interpret = "enochian_lm.translation.interpretation.cli:main"
enochian-apply-analytics = "enochian_lm.root_extraction.scripts.apply_analytics_updates:main"
enlm = "enochian_lm.analysis.cli:main"
